Move `recorded_logging_context.py` to `test_helpers`.
